Some moving.
git-svn-id: http://redline-rpm.googlecode.com/svn/redline/trunk@13 da7a95ae-83ad-11de-b471-8552911d2254

+++ source/main/org/freecompany/redline/header/Lead.java
@@ -18,11 +18,11 @@ public class Lead {
 	protected short os;
 	protected short sigtype;
 
-	public void setMajor( int major) {
+	public void setMajor( byte major) {
 		this.major = major;
 	}
 
-	public void setMinor( int minor) {
+	public void setMinor( byte minor) {
 		this.minor = minor;
 	}
